+++ alicloud/acctest/acctest.go
@@ -7,21 +7,21 @@ import (
 	"testing"
 
 	"github.com/aliyun/terraform-provider-alicloud/alicloud"
-	"github.com/aliyun/terraform-provider-alicloud/alicloud/provider/framework"
+	"github.com/aliyun/terraform-provider-alicloud/alicloud/provider"
 	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
 )
 
 // ProtoV5ProviderFactories serves the real muxed provider — the SDK v2 half and the
 // framework half behind one server — under the name given, defaulting to "alicloud".
 //
-// It cannot be called from a test that uses terraform-plugin-testing. This package
-// imports package alicloud for Provider(), and package alicloud imports
-// terraform-plugin-sdk/v2/helper/resource, which registers a -sweep flag in its init just
-// as terraform-plugin-testing/helper/resource does. A test binary linking both dies at
-// startup with "flag redefined: sweep" before any test runs. Until package alicloud moves
-// its ~1900 non-test files off helper/resource onto helper/retry, a test on the current
-// testing framework has to build its own factory; see protoV5ProviderFactories in
-// alicloud/function/arn_build_test.go for a framework-only one.
+// It works from a test on either harness, terraform-plugin-sdk/v2/helper/resource or
+// terraform-plugin-testing/helper/resource, but only because no non-test file in package
+// alicloud imports the former any more: both register a -sweep flag in their init, and a
+// test binary linking both dies at startup with "flag redefined: sweep" before any test
+// runs. This package imports package alicloud for Provider(), so a single non-test file
+// in package alicloud going back to helper/resource resurrects that panic for every
+// terraform-plugin-testing test that reaches this factory. Non-test code retries with
+// helper/retry instead.
 func ProtoV5ProviderFactories(names ...string) map[string]func() (tfprotov5.ProviderServer, error) {
 	if len(names) == 0 {
 		names = []string{"alicloud"}
@@ -30,7 +30,7 @@ func ProtoV5ProviderFactories(names ...string) map[string]func() (tfprotov5.Prov
 	factories := make(map[string]func() (tfprotov5.ProviderServer, error), len(names))
 	for _, name := range names {
 		factories[name] = func() (tfprotov5.ProviderServer, error) {
-			serverFactory, err := framework.ProtoV5ProviderServerFactory(context.Background(), alicloud.Provider())
+			serverFactory, err := provider.ProtoV5ProviderServerFactory(context.Background(), alicloud.Provider())
 			if err != nil {
 				return nil, err
 			}
