+++ alicloud/service/ims/data_source_default_domain_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/aliyun/terraform-provider-alicloud/alicloud/acctest"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
 
 func TestAccAliCloudImsDefaultDomainDataSource(t *testing.T) {
